refactors shutdown and uptime responses
